requirements.txt
anyio==3.4.0
asgiref==3.4.1
certifi==2020.6.20
click==8.0.3
colorama==0.4.4
contextlib2==21.6.0
contextvars==2.4
cycler==0.11.0
dataclasses==0.8
decorator==4.4.2
fastapi==0.71.0
Flask==2.0.2
h11==0.12.0
idna==3.3
imageio==2.9.0
immutables==0.16
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.3.4
networkx==2.5.1
numpy==1.19.5
opencv-python==4.5.5.62
opencv-python-headless==4.5.5.62
Pillow==8.2.0
pydantic==1.9.0
pyparsing==3.0.6
pytesseract==0.3.8
python-bidi==0.4.2
python-dateutil==2.8.2
python-multipart==0.0.5
PyWavelets==1.1.1
PyYAML==6.0
scikit-image==0.17.2
scipy==1.5.4
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
tifffile==2020.9.3
torch==1.10.1
torchvision==0.11.2
typing_extensions==4.0.1
uvicorn==0.16.0
Werkzeug==2.0.2
wincertstore==0.2
zipp==3.6.0
